+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Based on [apache/commons-compress@d0aec98](https://github.com/apache/commons-com
 * New interface `DataInputReadableChannel` for reading primitive data types from a `ReadableByteChannel`.
 * New interface `DataInputSeekableChannel` for reading primitive data types from a `SeekableByteChannel`.
 * New class `BufferedDataInputSeekableChannel` for buffering reads from a `SeekableByteChannel`.
+* Refactor `ZipArchiveReader` to use `DataInputSeekableChannel` instead of `SeekableByteChannel`.
 
 ## 1.27.1-1 (2024-12-01)